Add doc/, change Makefile and add some doc
The doc/ directory contain the configuration for doxygen and the doc.
Two functions were added to the Makefile, clean to clean all useless files and documentation to create the documentation.
The doc for the only function in network.c was added and the description of the game_mechanics and main files was added.