Adding gui.cpp, gui.h, media/ media/factorywars.png, changing Makefile
gui.h is the header of gui.cpp.
gui.cpp contain the code of the GUI, for now, it displays a png file on the screen and wait for the event SDL_QUIT.
media/ contains the media files.
media/factorywars.png is the image displayed by gui.cpp.
In the makefile, the target “all” now compile gui.cpp too.